# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/payouts": {
      "post": {
        "description": "Create a new fiat payout to a bank account",
        "operationId": "PayoutsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payout created successfully."
          },
          "400": {
            "description": "Validation failed or the request body/params were malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Create payout",
        "tags": [
          "Payouts"
        ]
      }
    },
    "/api/v1/payouts/resolve": {
      "post": {
        "description": "Verify and resolve bank account details to get account name",
        "operationId": "PayoutsController_resolveAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bank account resolved successfully."
          },
          "400": {
            "description": "Validation failed or the request body/params were malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Resolve bank account",
        "tags": [
          "Payouts"
        ]
      }
    },
    "/api/v1/payouts/{payout_id}": {
      "get": {
        "description": "Retrieve details of a specific payout by its ID",
        "operationId": "PayoutsController_findOne",
        "parameters": [
          {
            "name": "payout_id",
            "required": true,
            "in": "path",
            "description": "Payout unique identifier",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payout retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get payout by ID",
        "tags": [
          "Payouts"
        ]
      }
    },
...
    "/api/v1/misc/banks": {
      "get": {
        "operationId": "MiscController_getBanks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Supported banks retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get supported banks",
        "tags": [
          "Misc - Banks & Assets"
        ]
      },
      "post": {
        "operationId": "MiscController_addBank",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBankDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bank account added successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "The resource already exists or is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Add bank account",
        "tags": [
          "Misc - Banks & Assets"
        ]
      }
    },
    "/api/v1/misc/banks/saved": {
      "get": {
        "operationId": "MiscController_getSavedBanks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Saved banks retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get saved banks",
        "tags": [
          "Misc - Banks & Assets"
        ]
      }
    },
    "/api/v1/misc/banks/saved/{accountNumber}": {
      "get": {
        "operationId": "MiscController_getSavedBankById",
        "parameters": [
          {
            "name": "accountNumber",
            "required": true,
            "in": "path",
            "schema": {
              "example": "3154021148",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Saved bank retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get saved bank by account number",
        "tags": [
          "Misc - Banks & Assets"
        ]
      }
    },
    "/api/v1/misc/banks/{bankId}": {
      "delete": {
        "operationId": "MiscController_deleteBank",
        "parameters": [
          {
            "name": "bankId",
            "required": true,
            "in": "path",
            "description": "Local bank record ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bank account deleted successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Delete bank account",
        "tags": [
          "Misc - Banks & Assets"
        ]
      }
    },
    "/api/v1/misc/banks/verify": {
      "post": {
        "operationId": "MiscController_verifyBankAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyBankAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bank account verified successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify bank account",
        "tags": [
          "Misc - Banks & Assets"
        ]
      }
    },
...
    "/api/v1/misc/crypto-prices": {
      "post": {
        "operationId": "MiscController_getCryptoPrices",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetRateQueryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Crypto conversion rate retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get crypto conversion rate",
        "tags": [
          "Misc - Banks & Assets"
        ]
      }
    },
    "/api/v1/misc/rate-calculator": {
      "post": {
        "operationId": "MiscController_getRateCalculator",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "asset": {
                    "type": "string",
                    "example": "USDT"
                  },
                  "amount": {
                    "type": "number",
                    "example": 100
                  },
                  "currency": {
                    "type": "string",
                    "example": "NGN"
                  }
                },
                "required": [
                  "asset",
                  "amount",
                  "currency"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Estimated payment amount calculated successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Estimate payment amount in crypto",
        "tags": [
          "Misc - Banks & Assets"
        ]
      }
    },
...
    "/api/v1/invoices/pay/{invoice_number}": {
      "get": {
        "description": "Public endpoint \u2014 no auth required. Used by the customer-facing pay page to display invoice details.",
        "operationId": "InvoicesController_findByNumber",
        "parameters": [
          {
            "name": "invoice_number",
            "required": true,
            "in": "path",
            "description": "Invoice number e.g. INV-ABC123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice retrieved successfully."
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get invoice by number (public)",
        "tags": [
          "Invoices"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dexxify API",
    "description": "Crypto Infrastructure API for Africa (Wallets, Payouts, Offramp, Onramp, KYC, KYB)",
    "version": "1.0",
    "contact": {
      "name": "Dexxify",
      "url": "https://www.dexxify.com",
      "email": "dexxifyhq@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "https://api.dexxify.com/api/v1",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "API key for /api/v1/* endpoints"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed."
          },
          "errors": {
            "type": "object",
            "example": [
              "amount must not be less than 0"
            ]
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-17T12:00:00.000Z"
          }
        },
        "required": [
          "success",
          "status",
          "message",
          "timestamp"
        ]
      },
...
      "CreatePayoutDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Payout amount in NGN",
            "example": 5000,
            "minimum": 100
          },
          "bank_code": {
            "type": "string",
            "description": "Bank code",
            "example": "057"
          },
          "account_number": {
            "type": "string",
            "description": "Account number",
            "example": "1234567890"
          },
          "account_name": {
            "type": "string",
            "description": "Account name",
            "example": "John Doe"
          },
          "narration": {
            "type": "string",
            "description": "Payment narration",
            "example": "Salary payment"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata",
            "example": {
              "source": "payroll"
            }
          }
        },
        "required": [
          "amount",
          "bank_code",
          "account_number"
        ]
      },
      "ResolveAccountDto": {
        "type": "object",
        "properties": {
          "bank_code": {
            "type": "string",
            "description": "Bank code",
            "example": "057"
          },
          "account_number": {
            "type": "string",
            "description": "Account number",
            "example": "1234567890"
          }
        },
        "required": [
          "bank_code",
          "account_number"
        ]
      },
...
      "AddBankDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string",
            "description": "Bank account number",
            "example": "8065924354"
          },
          "bankCode": {
            "type": "string",
            "description": "Bank code (e.g. 305 for OPay)",
            "example": "305"
          },
          "label": {
            "type": "string",
            "description": "Friendly label",
            "example": "My OpayBank"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Default account recipient",
            "example": false
          }
        },
        "required": [
          "accountNumber",
          "bankCode",
          "isDefault"
        ]
      },
      "VerifyBankAccountDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string",
            "description": "Bank account number",
            "example": "2249098732"
          },
          "bankCode": {
            "type": "string",
            "description": "Bank code (e.g. 033 for UBABank)",
            "example": "033"
          }
        },
        "required": [
          "accountNumber",
          "bankCode"
        ]
      },
      "GetRateQueryDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "description": "Source asset",
            "example": "USDT"
          },
          "to": {
            "type": "string",
            "description": "Fiat currency",
            "example": "NGN"
          }
        },
        "required": [
          "from",
          "to"
        ]
      },
...
    }
  }
}